Conflicts:
	CMakeLists.txt
	intern/cycles/device/device_opencl.cpp
	intern/cycles/kernel/geom/geom_motion_triangle.h
	intern/cycles/kernel/kernel_light.h
	intern/cycles/kernel/kernel_passes.h
	intern/cycles/kernel/kernel_path.h
	intern/cycles/kernel/kernel_shader.h
	intern/cycles/kernel/kernel_volume.h
	release/datafiles/fonts/droidsans.ttf.gz
	source/blender/blenkernel/BKE_curve.h
	source/blender/blenkernel/intern/image.c
	source/blender/blenkernel/intern/material.c
	source/blender/blenkernel/intern/mesh_remap.c
	source/blender/blenkernel/intern/sequencer.c
	source/blender/blenloader/intern/versioning_270.c
	source/blender/blenloader/intern/writefile.c
	source/blender/bmesh/intern/bmesh_core.c
	source/blender/bmesh/intern/bmesh_core.h
	source/blender/bmesh/intern/bmesh_operator_api_inline.h
	source/blender/bmesh/intern/bmesh_queries.c
	source/blender/compositor/nodes/COM_ImageNode.cpp
	source/blender/compositor/operations/COM_BlurBaseOperation.cpp
	source/blender/compositor/operations/COM_CompositorOperation.cpp
	source/blender/editors/include/ED_screen.h
	source/blender/editors/interface/interface_templates.c
	source/blender/editors/mesh/editmesh_rip.c
	source/blender/editors/mesh/editmesh_select.c
	source/blender/editors/render/render_internal.c
	source/blender/editors/render/render_opengl.c
	source/blender/editors/render/render_preview.c
	source/blender/editors/screen/area.c
	source/blender/editors/sculpt_paint/paint_image_proj.c
	source/blender/editors/space_file/filelist.c
	source/blender/editors/space_image/image_buttons.c
	source/blender/editors/space_image/image_draw.c
	source/blender/editors/space_sequencer/sequencer_draw.c
	source/blender/editors/space_view3d/view3d_draw.c
	source/blender/editors/space_view3d/view3d_ruler.c
	source/blender/render/extern/include/RE_pipeline.h
	source/blender/render/intern/source/pipeline.c
	source/blender/render/intern/source/render_result.c
	source/blenderplayer/bad_level_call_stubs/stubs.c

+++ intern/cycles/device/device_opencl.cpp
@@ -858,6 +858,7 @@ public:
 
 		if(debug_src)
 			path_write_text(*debug_src, source);
+
 		size_t source_len = source.size();
 		const char *source_str = source.c_str();
 
@@ -3398,7 +3399,7 @@ public:
 	*/
 	int2 get_max_render_feasible_tile_size(size_t feasible_global_work_size) {
 		int2 max_render_feasible_tile_size;
-		int square_root_val = sqrt(feasible_global_work_size);
+		int square_root_val = (int)sqrt(feasible_global_work_size);
 		max_render_feasible_tile_size.x = square_root_val;
 		max_render_feasible_tile_size.y = square_root_val;
